requirements.txt
click
clint
logzero
profig
psutil
pycrypto
hot-redis
paho-mqtt
wcwidth